Refactor login and registration handlers
